requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.3
djangorestframework==3.11.0
Faker==4.1.0
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
